refactor(documents-to-convert): Add documents-to-convert.model
